# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trainerplan API
  version: 1.0.0
  description: |
    Versionierte HTTP-API für die Trainerplan-Web-App und native Clients.
    Browser verwenden eine HTTP-only-Cookie-Session; native Clients verwenden
    das beim Login ausgegebene Bearer-Token.
servers:
  - url: /api/v1
    description: Aktuelle API-Version
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /auth/register/club:
    post:
      tags: [Authentifizierung]
      security: []
      summary: Verein mit erster Mannschaft und Admin-Konto anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClubRegistrationRequest' }
      responses:
        '201':
          description: Verein, Mannschaft und Admin wurden angelegt
          content:
            application/json:
              schema:
                type: object
                required: [user, club, token, expiresAt]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  club: { $ref: '#/components/schemas/ClubRegistrationResult' }
                  token: { type: string }
                  expiresAt: { type: string, format: date-time }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register:
    post:
      tags: [Authentifizierung]
      security: []
      summary: Persönlichen Zugang anlegen
      description: Erstellt nur den Zugang. Vereins- und Profildaten werden nach der ersten Anmeldung über das Onboarding erfasst.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12, maxLength: 256 }
      responses:
        '201': { description: Zugang erstellt }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /onboarding/club:
    post:
      tags: [Authentifizierung]
      summary: Verein nach der ersten Anmeldung einrichten
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, clubName, ageGroup]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                clubName: { type: string, minLength: 2, maxLength: 120 }
                designation: { type: string, maxLength: 80, description: Optionale Zusatzbezeichnung, z. B. Blau oder Mädchen }
                ageGroup: { type: string, enum: [g2, g1, f2, f1, e2, e1, d2, d1, c2, c1, b2, b1, a2, a1] }
      responses:
        '200': { description: Verein eingerichtet }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /events:
    get:
      tags: [Termine]
      summary: Sichtbare Termine laden
      responses:
        '200':
          description: Termine
          content:
            application/json:
              schema:
                type: object
                properties: { events: { type: array, items: { $ref: '#/components/schemas/Event' } } }
    post:
      tags: [Termine]
      summary: Termin erstellen
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EventWrite' } } }
      responses:
        '201':
          description: Erstellter Termin
          content: { application/json: { schema: { type: object, properties: { event: { $ref: '#/components/schemas/Event' } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Termine]
      summary: Terminliste oder Teilnahmeänderungen speichern (Kompatibilität)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties: { events: { type: array, items: { $ref: '#/components/schemas/Event' } } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /events/{id}:
    parameters:
      - { $ref: '#/components/parameters/ResourceId' }
    patch:
      tags: [Termine]
      summary: Termin ändern
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EventWrite' } } }
      responses:
        '200': { description: Termin geändert, content: { application/json: { schema: { type: object, properties: { event: { $ref: '#/components/schemas/Event' } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Termine]
      summary: Termin löschen
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /events/{id}/attendance:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    put:
      tags: [Termine]
      summary: Eigene Teilnahme setzen oder entfernen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { value: { oneOf: [{ $ref: '#/components/schemas/Attendance' }, { type: 'null' }] } }
      responses:
        '200': { description: Teilnahme aktualisiert, content: { application/json: { schema: { type: object, properties: { event: { $ref: '#/components/schemas/Event' } } } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /events/{id}/squads:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    get:
      tags: [Mannschaft]
      summary: Turniermannschaften laden
      responses:
        '200': { description: Mannschaftsplanung geladen, content: { application/json: { schema: { $ref: '#/components/schemas/SquadsPayload' } } } }
    put:
      tags: [Mannschaft]
      summary: Turniermannschaften speichern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [squads]
              properties: { squads: { type: array, items: { $ref: '#/components/schemas/Squad' } } }
      responses:
        '200': { description: Mannschaftsplanung gespeichert, content: { application/json: { schema: { $ref: '#/components/schemas/SquadsPayload' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: session-token }
    sessionCookie: { type: apiKey, in: cookie, name: trainerplan_session }
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
  responses:
    Ok:
      description: Erfolgreich
      content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } }
...
    Unauthorized:
      description: Nicht angemeldet
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Keine Berechtigung
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Ressource nicht gefunden
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Konflikt
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Zu viele Anfragen
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties: { error: { type: string } }
    Ok:
      type: object
      properties: { ok: { type: boolean, example: true } }
    Role: { type: string, enum: [admin, trainer, player] }
    Attendance: { type: string, enum: [yes, no, maybe] }
...
    ClubRegistrationRequest:
      type: object
      required: [name, email, password, clubName, teamName]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 256 }
        clubName: { type: string, minLength: 2, maxLength: 120 }
        teamName: { type: string, minLength: 2, maxLength: 120 }
        ageGroup: { type: string, maxLength: 40 }
    ClubRegistrationResult:
      type: object
      required: [name, teamName]
      properties:
        name: { type: string }
        teamName: { type: string }
    User:
      type: object
      required: [id, name, email, role, position, phone, birthday, ageGroup, dribblingRating, shootingRating, passingRating]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        position: { type: string }
        number: { type: [integer, 'null'] }
        ballNumber: { type: [integer, 'null'] }
        phone: { type: string }
        birthday: { type: string, format: date }
        ageGroup: { type: string }
        avatar: { type: [string, 'null'] }
        groupId: { type: [string, 'null'] }
        dribblingRating: { type: integer, minimum: 0, maximum: 5 }
        shootingRating: { type: integer, minimum: 0, maximum: 5 }
        passingRating: { type: integer, minimum: 0, maximum: 5 }
        internalTeam: { type: [string, 'null'], enum: [A, B, null] }
    EventWrite:
      type: object
      required: [type, title, date, startTime, endTime, meetingTime, location, description, maxParticipants]
      properties:
        type: { type: string, enum: [training, tournament, event] }
        title: { type: string }
        date: { type: string, format: date }
        startTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endTime: { type: string }
        meetingTime: { type: string }
        location: { type: string }
        address: { type: string }
        description: { type: string }
        trainerNote: { type: string }
        weather: { $ref: '#/components/schemas/Weather' }
        maxParticipants: { type: integer, minimum: 1 }
    Event:
      allOf:
        - $ref: '#/components/schemas/EventWrite'
        - type: object
          required: [id, responses]
          properties:
            id: { type: string }
            responses: { type: object, additionalProperties: { $ref: '#/components/schemas/Attendance' } }
    Weather:
      type: object
      required: [condition, label, temperature]
      properties:
        condition: { type: string, enum: [sunny, partly-cloudy, cloudy] }
        label: { type: string }
        temperature: { type: integer, minimum: -50, maximum: 60 }
...
    Squad:
      type: object
      required: [id, eventId, name, playerIds]
      properties:
        id: { type: string }
        eventId: { type: string }
        name: { type: string }
        trainerId: { type: [string, 'null'] }
        playerIds: { type: array, items: { type: string } }
    SquadsPayload:
      type: object
      required: [eventId, squads]
      properties:
        eventId: { type: string }
        squads: { type: array, items: { $ref: '#/components/schemas/Squad' } }
...